requirements.txt
aiogram
APScheduler
python-decouple
requests
gspread
